// src/types/Units.d.ts
export type AbsoluteUnits = 'px' | 'cm' | 'mm' | 'Q' | 'in' | 'pt' | 'pc';
export type RelativeUnits = 'em' | 'ex' | 'ch' | 'rem' | 'vw' | 'vh' | 'vmin' | 'vmax' | '%';
export type DynamicUnits = 'dvw' | 'dvh' | 'dvmin' | 'dvmax';
export type StaticUnits = 'cm' | 'mm' | 'Q' | 'in' | 'px' | 'pt' | 'pc' | 'em' | 'rem';
// ...
export type StandardUnits = AbsoluteUnits | RelativeUnits | DynamicUnits;
// ...
export type AbsoluteSize = `${number}${AbsoluteUnits}`;
export type RelativeSize = `${number}${RelativeUnits}`;
export type DynamicSize = `${number}${DynamicUnits}`;
export type StaticSize = `${number}${StaticUnits}`;
export type ExtRelativeSize = `${number}${ExtendedRelativeUnits}`;
export type StandardViewportSize = `${number}${StandardViewportUnits}`;
export type LargeViewportSize = `${number}${LargeViewportUnits}`;
export type SmallViewportSize = `${number}${SmallViewportUnits}`;
export type DynamicViewportSize = `${number}${DynamicViewportUnits}`;
export type StandardSize = AbsoluteSize | RelativeSize | DynamicSize;